Using Supplier interface instead of the custom LazyInitializer
